# Non-animated property change notifications