--- a/src/Avalonia.Base/PropertyStore/ConstantValueEntry.cs
+++ b/src/Avalonia.Base/PropertyStore/ConstantValueEntry.cs
@@ -13,6 +13,7 @@ namespace Avalonia.PropertyStore
     internal class ConstantValueEntry<T> : IPriorityValueEntry<T>, IDisposable
     {
         private IValueSink _sink;
+        private Optional<T> _value;
 
         public ConstantValueEntry(
             StyledPropertyBase<T> property,
@@ -21,18 +22,21 @@ namespace Avalonia.PropertyStore
             IValueSink sink)
         {
             Property = property;
-            Value = value;
+            _value = value;
             Priority = priority;
             _sink = sink;
         }
 
         public StyledPropertyBase<T> Property { get; }
         public BindingPriority Priority { get; }
-        public Optional<T> Value { get; }
-        Optional<object> IValue.Value => Value.ToObject();
-        BindingPriority IValue.ValuePriority => Priority;
+        Optional<object> IValue.GetValue() => _value.ToObject();
 
-        public void Dispose() => _sink.Completed(Property, this, Value);
+        public Optional<T> GetValue(BindingPriority maxPriority = BindingPriority.Animation)
+        {
+            return Priority >= maxPriority ? _value : Optional<T>.Empty;
+        }
+
+        public void Dispose() => _sink.Completed(Property, this, _value);
         public void Reparent(IValueSink sink) => _sink = sink;
     }
 }