--- a/src/Avalonia.Base/PropertyStore/LocalValueEntry.cs
+++ b/src/Avalonia.Base/PropertyStore/LocalValueEntry.cs
@@ -14,9 +14,14 @@ namespace Avalonia.PropertyStore
         private T _value;
 
         public LocalValueEntry(T value) => _value = value;
-        public Optional<T> Value => _value;
-        public BindingPriority ValuePriority => BindingPriority.LocalValue;
-        Optional<object> IValue.Value => Value.ToObject();
+        public BindingPriority Priority => BindingPriority.LocalValue;
+        Optional<object> IValue.GetValue() => new Optional<object>(_value);
+        
+        public Optional<T> GetValue(BindingPriority maxPriority)
+        {
+            return BindingPriority.LocalValue >= maxPriority ? _value : Optional<T>.Empty;
+        }
+
         public void SetValue(T value) => _value = value;
     }
 }