… src/Avalonia.Base/PropertyStore/PriorityValue.cs
… src/Avalonia.Base/ValueStore.cs
… src/Avalonia.Controls/Button.cs
… src/Avalonia.Controls/ButtonSpinner.cs
… src/Avalonia.Controls/Calendar/DatePicker.cs
… src/Avalonia.Controls/Expander.cs
… src/Avalonia.Controls/Notifications/WindowNotificationManager.cs
… src/Avalonia.Controls/Primitives/ScrollBar.cs
… src/Avalonia.Controls/Primitives/Track.cs
… src/Avalonia.Controls/ProgressBar.cs
… src/Avalonia.Controls/Repeater/ItemsRepeater.cs
… src/Avalonia.Controls/Slider.cs
… src/Avalonia.Controls/Window.cs
… src/Avalonia.Input/InputElement.cs
… src/Avalonia.Layout/StackLayout.cs
… src/Avalonia.Layout/UniformGridLayout.cs
… src/Avalonia.Styling/Styling/PropertySetterBindingInstance.cs
… src/Avalonia.Visuals/Media/DrawingImage.cs
… tests/Avalonia.Base.UnitTests/AvaloniaObjectTests_GetValue.cs
… tests/Avalonia.Base.UnitTests/AvaloniaObjectTests_OnPropertyChanged.cs
… tests/Avalonia.Base.UnitTests/AvaloniaPropertyTests.cs
… tests/Avalonia.Base.UnitTests/PriorityValueTests.cs
… tests/Avalonia.Markup.Xaml.UnitTests/Xaml/InitializationOrderTracker.cs
… tests/Avalonia.Visuals.UnitTests/Rendering/ImmediateRendererTests.cs