## What does the pull request do?

As described in #3109, transitions were not working correctly because there was no way to get change notifications for non-animated property changes.

This PR is the first part in fixing #3109 - it does NOT fix transitions, but it does provide a way to get property change notifications for non-animated property changes. Transitions are fixed in #3854 

- Pass information for all property changes to `OnPropertyChangedCore`; whether they result in an effective value change or not
- `AvaloniaPropertyChangedEventArgs` now has an `IsEffectiveValue` property which will be false when a higher priority value is active
- `OnPropertyChangedCore` calls `OnPropertyChanged` with all notifications which have `IsEffectiveValue == true`
- Added `AvaloniaObject.GetBaseValue` to get a value with a specified priority
- Change the signature of `OnPropertyChanged` again (sorry) to take an `AvaloniaPropertyChangedEventArgs<T>`

## Checklist

- [x] Added unit tests (if possible)?

## Breaking changes

The signature for `OnPropertyChanged` has changed again.