@MarchingCube 2020-05-01
In case somebody wonders about perf impact:

This PR
|         Method |        Mean |       Error |      StdDev |    Gen 0 |    Gen 1 |   Gen 2 |  Allocated |
|--------------- |------------:|------------:|------------:|---------:|---------:|--------:|-----------:|
| CreateCalendar | 20,164.0 us | 224.2566 us | 209.7698 us | 687.5000 | 281.2500 | 31.2500 | 3962.71 KB |
|   CreateButton |    139.0 us |   0.9656 us |   0.8063 us |   6.8359 |   0.9766 |       - |   28.66 KB |
|  CreateTextBox |  1,080.0 us |   4.4062 us |   3.6794 us |  42.9688 |  21.4844 |       - |  183.04 KB |

Master
|         Method |        Mean |      Error |     StdDev |    Gen 0 |    Gen 1 | Gen 2 |  Allocated |
|--------------- |------------:|-----------:|-----------:|---------:|---------:|------:|-----------:|
| CreateCalendar | 19,973.4 us | 232.356 us | 194.027 us | 593.7500 | 250.0000 |     - | 3630.23 KB |
|   CreateButton |    140.4 us |   2.530 us |   2.366 us |   6.3477 |   0.9766 |     - |   26.77 KB |
|  CreateTextBox |  1,114.1 us |  19.217 us |  17.976 us |  41.0156 |  19.5313 |     - |  172.22 KB |

So perf is withing margin of error and we now need roughly 10% more memory.