@Gillibald 2020-05-01
Is that extra memory consumption coming from the event args?