@MarchingCube 2020-05-01
> Is that extra memory consumption coming from the event args?

I think so, but I didn't run memory profiler on it so just a guess.