@grokys 2020-05-03
>> Is that extra memory consumption coming from the event args?
> I think so, but I didn't run memory profiler on it so just a guess.

Yeah, quite possibly: we're now doing more property changed notifications. A further improvement will be to pool `AvaloniaPropertyChangedEventArgs<T>` objects.